Domain conversion to objectify round 1
Former-commit-id: a1d09c93c4931e4adb78fbce4f3548804669432e

+++ gwtp-sample-carstore/src/main/java/com/arcbees/carsample/client/application/manufacturer/ManufacturerDetailPresenter.java
@@ -81,7 +81,7 @@ public class ManufacturerDetailPresenter extends Presenter<MyView, MyProxy> impl
         createNew = Strings.isNullOrEmpty(param);
 
         if (!createNew) {
-            Integer id = Integer.parseInt(param);
+            Long id = Long.parseLong(param);
             dispatcher.execute(new GetManufacturerAction(id), new SafeAsyncCallback<GetResult<Manufacturer>>() {
                 @Override
                 public void onSuccess(GetResult<Manufacturer> result) {
